add separate event for citizen payment
